refactor: improve import file structure
